+++ packages/restapi/src/lib/constants.ts
@@ -10,10 +10,10 @@ const Constants = {
     LIMIT_MIN: 1,
     LIMIT_MAX: 50
   },
-  DEFAULT_CHAIN_ID: 42,
+  DEFAULT_CHAIN_ID: 5,
   DEV_CHAIN_ID: 99999,
   NON_ETH_CHAINS: [137, 80001],
-  ETH_CHAINS: [1, 42]
+  ETH_CHAINS: [1, 5]
 };
 
 export default Constants;